add request body validation

+++ src/main/java/com/example/dctesttask/department/DepartmentController.java
@@ -1,7 +1,8 @@
 package com.example.dctesttask.department;
 
+import com.example.dctesttask.employee.EmployeeRepository;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -13,6 +14,7 @@ import static org.springframework.http.HttpStatus.*;
 public class DepartmentController {
 
     private final DepartmentRepository departmentRepository;
+    private final EmployeeRepository employeeRepository;
 
     @GetMapping
     public Iterable<Department> findAll() {
@@ -27,17 +29,21 @@ public class DepartmentController {
     @DeleteMapping("/{id}")
     @ResponseStatus(NO_CONTENT)
     public void deleteById(@PathVariable Long id) {
+        if (employeeRepository.existsByDepartmentId(id)) {
+            throw new ResponseStatusException(BAD_REQUEST, "Can't delete department with employees attached");
+        }
+
         departmentRepository.deleteById(id);
     }
 
     @PostMapping
     @ResponseStatus(CREATED)
-    public Department create(@RequestBody Department department) {
+    public Department create(@Valid @RequestBody Department department) {
         return departmentRepository.save(department);
     }
 
     @PutMapping("/{id}")
-    public Department update(@PathVariable Long id, @RequestBody Department department) {
+    public Department update(@PathVariable Long id, @Valid @RequestBody Department department) {
         var existingDep = verifyDepartment(id);
 
         existingDep.setName(department.getName());
